requirements.txt
pyupbit
python-telegram-bot==13.15
